+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2020-07-14
+
+### ✨ OptionsPanel
+
+### Added
+
+- `<OptionsPanel />` component
+- `<Checkbox />` component
+- `<Input />` component
+
+### Changed
+
+### Removed
+
+### Fixed
+
 ## [0.11.0] - 2020-07-14
 
 ### 📄 Job Card